very minor refactor
